pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "remote-claude-mcp"
version = "1.0.0"
description = "MCP server that proxies Claude Code tools to remote clusters via SSH"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
remote-claude-mcp = "remote_claude_mcp.__main__:main"
